+++ unicorn_binance_websocket_api/unicorn_binance_websocket_api_connection.py
@@ -99,6 +99,24 @@ class BinanceWebSocketApiConnection(object):
                     del self.handler_binance_websocket_api_manager.restart_requests[self.stream_id]
             except KeyError:
                 pass
+        except ConnectionResetError as error_msg:
+            logging.critical("BinanceWebSocketApiConnection->await._conn.__aenter__(" + str(self.stream_id) + ", " +
+                             str(self.channels) + ", " + str(self.markets) + ")" + " - ConnectionResetError "
+                             ";) - " + str(error_msg))
+            self.handler_binance_websocket_api_manager.stream_is_crashing(self.stream_id, (str(error_msg) +
+                                                                          " - ConnectionResetError"))
+            time.sleep(0.5)
+            self.handler_binance_websocket_api_manager.set_restart_request(self.stream_id)
+            sys.exit(1)
+        except OSError as error_msg:
+            logging.critical("BinanceWebSocketApiConnection->await._conn.__aenter__(" + str(self.stream_id) + ", " +
+                             str(self.channels) + ", " + str(self.markets) + ")" + " - OSError "
+                             ";) - " + str(error_msg))
+            self.handler_binance_websocket_api_manager.stream_is_crashing(self.stream_id, (str(error_msg) +
+                                                                          " - OSError"))
+            time.sleep(0.5)
+            self.handler_binance_websocket_api_manager.set_restart_request(self.stream_id)
+            sys.exit(1)
         except socket.gaierror as error_msg:
             logging.critical("BinanceWebSocketApiConnection->await._conn.__aenter__(" + str(self.stream_id) + ", " +
                              str(self.channels) + ", " + str(self.markets) + ")" + " - No internet connection? "